Fix for some MSVC compile problems and warnings.

+++ src/gui/components/lookandfeel/juce_LookAndFeel.cpp
@@ -470,7 +470,7 @@ void LookAndFeel::drawToggleButton (Graphics& g,
     if (! button.isEnabled())
         g.setOpacity (0.5f);
 
-    const int textX = tickWidth + 5;
+    const int textX = (int) tickWidth + 5;
 
     g.drawFittedText (button.getButtonText(),
                       textX, 0,
